Sample Data 1: senders.json
This file represents the "email sender" data. The full file will contain an array of objects, each object representing a possible email sender. For now, each object will contain the emailer's name, job title, color of their name square, and an "id" attribute that we'll use to link this data to a particular email in the questions.json file.