Sample Data 3: dialog-tree.json
This data structure represents the decision tree for the game. "trees" contains an array of objects. Each object's "scenario_id" connects to the "id" defined in scenario.json. The "branches" then show how we flow through the game. The "from" and "to" fields connect to question IDs in question.json. The "criteria" field indicates what would cause you to go from one question to the next - assume that in the future, it will be a comma-separated field of possible values.