Sample Data 4: questions.json
This is the big one. "scenario_id" again matches with the ID attribute in "scenario.json". The "questions" attribute contains an array of question objects. Each object has:
* Its own ID (referenced in dialog-tree.json)
* A "sender_id" to connect it to the senders.json data
* Email subject line and question text
* An array of answer objects that contain the text of the answer and any criteria they set (again, connected to the dialog-tree.json)

As a suggestion, make a dictionary and use it to set values from criteria (e.g. dict["q1_correct"] = "0"). In the future, we might be checking multiple criteria in dialog-tree.json in order to determine which question to load next.