Almost complete Pre-Assessment Update
10 hours of work:

I got most of the Pre-assessment done except for the very last part which is where the user adjusted values of the attributes are saved back into the JSON and then proceeds to the next scene.

I will work on this part later this coming week and after that the first two scenes should be complete and ready for the prototype.